warning free

+++ src/xyc2msh.d/count_nodes.c++
@@ -6,7 +6,7 @@ using namespace std;
 int count_nodes(vector<nde>&N)
 {
   int n=0, i;
-  for (i=1; i<=N.size()-1; i++)
+  for (i=1; i<=(int)(N.size()-1); i++)
     if (N[i].a!=0) n++;
   return n;
 }
